% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltimore_plans}
\alias{baltimore_plans}
\title{Baltimore City and Regional Plans}
\format{
A data frame with 366 rows and 36 variables:
\describe{
\item{\code{title}}{Plan title}
\item{\code{year_complete}}{Year adopted/completed}
\item{\code{year_update}}{Year updated}
\item{\code{date_complete}}{Date (may be approximate)}
\item{\code{status}}{Current status}
\item{\code{document_url}}{Document URL}
\item{\code{document_filename}}{Document URL}
\item{\code{document_cover_filename}}{Document URL}
\item{\code{site_url}}{Related site URL}
\item{\code{geography}}{Geography}
\item{\code{document_type}}{Document type}
\item{\code{program}}{Related program}
\item{\code{prepared_for}}{Entity or entities prepared for}
\item{\code{prepared_by}}{Entity or entities prepared by}
\item{\code{funded_by}}{Entity or entities funded by}
\item{\code{location_names}}{Location names}
\item{\code{county}}{County name}
\item{\code{main_topic}}{Main topic}
\item{\code{topics}}{Topics}
\item{\code{name_abb}}{Name abbreviation}
\item{\code{name_short}}{Short name}
\item{\code{title_alt}}{Alternate title for spatial data join}
\item{\code{short_description}}{Short document description}
\item{\code{document_page_count}}{Page count}
\item{\code{document_url_type}}{Page count}
\item{\code{status_organization}}{Entity assigning status}
\item{\code{year_status}}{Status assignment date}
\item{\code{date_status}}{Status assignment date}
\item{\code{status_regulatory}}{Regulatory status}
\item{\code{source}}{Document source}
\item{\code{related_plans}}{Related plans (typically a preceding plan in a sequence of related plans)}
\item{\code{notes}}{Notes}
\item{\code{created_date}}{Airtable record created date/time}
\item{\code{last_modified}}{Airtable record last modified date/time}
\item{\code{record_id}}{Airtable record identifier}
\item{\code{plan_area_desc}}{Plan area description}
}
}
\source{
\url{https://airtable.com/appC0OOaNJtKU6iHz/shrIb0577rFcNI6vj}
}
\usage{
baltimore_plans
}
\description{
Table from the Baltimore City and Regional Plans Airtable base:
https://airtable.com/appC0OOaNJtKU6iHz/shrIb0577rFcNI6vj
}
\details{
This base is an index for community plans, park plans, transportation plans,
and related reports or studies for Baltimore City, adjacent counties, and the
Baltimore metro area.

The goal is to provide a convenient and accessible reference to information
plans include: what the scope and related locations for the plan, when the
plan was published or adopted, who prepared the plan, and any programs
connected to the plan. This data is the \strong{plans} table including plans,
studies, reports, etc.

Last updated: 2025-05-22
}
\keyword{datasets}
