% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baltimoredata-package.R
\docType{package}
\name{baltimoredata-package}
\alias{baltimoredata}
\alias{baltimoredata-package}
\title{baltimoredata: Access Updated and Historic Data on Baltimore City, Maryland}
\description{
This package provides access to a small number of spatial and nonspatial datasets related to Baltimore City, Maryland.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elipousson.github.io/baltimoredata/}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID})

}
\keyword{internal}
