% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa_atlas_maps}
\alias{msa_atlas_maps}
\title{Maryland State Archives Atlas Maps Index}
\format{
A data frame with 786 rows and 17 variables:
\describe{
\item{\code{source_id}}{MSA identifier}
\item{\code{title}}{Plate or page title}
\item{\code{type}}{Plate or page type}
\item{\code{sheet_title}}{Sheet title}
\item{\code{atlas}}{Atlas reference}
\item{\code{file_number}}{File number}
\item{\code{html_url}}{URL for HTML gallery page}
\item{\code{pdf_url}}{URL for PDF file}
\item{\code{filename}}{Filename}
\item{\code{year}}{Year (from atlas)}
\item{\code{source}}{Source (from atlas)}
\item{\code{description}}{Description}
\item{\code{locations}}{Locations (incomplete)}
\item{\code{page_number}}{Page number}
\item{\code{gif_url}}{URL for GIF format file}
\item{\code{pdf_right_url}}{URL for PDF file (right side)}
\item{\code{pdf_left_url}}{URL for PDF file (left side)}
}
}
\usage{
msa_atlas_maps
}
\description{
An index of digitized atlas maps available from the Maryland State Archives.
}
\details{
\url{https://airtable.com/appLdXzqz0QOXqJTE/shr8NPkZnJzBLjH9r}
}
\keyword{datasets}
