% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vital_signs_indicators}
\alias{vital_signs_indicators}
\title{BNIA Vital Signs Indicators, 2010-2018}
\format{
A data frame with 177 rows and 7 variables:
\describe{
  \item{\code{label}}{Indicator label}
  \item{\code{section}}{Vital Signs report section}
  \item{\code{source}}{Indicator data source}
  \item{\code{normalization_source}}{Normalization data source, e.g. Census data for CSA residents}
  \item{\code{definition}}{Indicator definition}
  \item{\code{indicator}}{Indicator variable name}
  \item{\code{years_available}}{Years available}
}
}
\usage{
vital_signs_indicators
}
\description{
Labels, definitions, and data sources for the Vital Signs
  indicators used by the Baltimore Neighborhood Indicators Alliance from 2010
  to 2018.
}
\keyword{datasets}
