% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity_xwalk}
\alias{entity_xwalk}
\title{Baltimore City Entity Crosswalk}
\format{
A data frame with 502 rows and 5 variables:
\describe{
\item{\code{id}}{Identifier ID}
\item{\code{name}}{Identifier name}
\item{\code{entity}}{Entity name}
\item{\code{source}}{Crosswalk data source}
\item{\code{use}}{Usage status}
}
}
\usage{
entity_xwalk
}
\description{
Note, that the "Baltimore_City_Employee_Residency_by_Agency" source includes
identifiers and name-only identifiers that match data found in the
"EmployeeSalaries_1" source.
}
\details{
Last updated: 2025-05-22
}
\keyword{datasets}
